config: add logtype of 'none' to discard all logging
Certain distros (like Fedora) discard Razor2::Agent::Client logging in
a round-about way (like creating a file which is a symlink to /dev/null)
but this might create SElinux issues in turn.

Provide a more direct way to discard logging which isn't vulnerable
to bizarre SElinux AVC's when walking paths to symlinks which just
point to /dev/null.

Signed-off-by: Philip Prindeville <philipp@cpan.org>